% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_web.R
\name{food_web}
\alias{food_web}
\title{Run a food web.}
\usage{
food_web(tmax, a, b, r, w, d, tstep = 1, V = TRUE, R = TRUE,
  H = TRUE, pool_starts = NULL)
}
\arguments{
\item{tmax}{Duration over which to run the model.}

\item{a}{Controls the smoothness of the pulse, along with \code{r}. If \code{a} is
sufficiently high, the pulse will always be rectangular.}

\item{b}{Maximum value of the midge pulse.}

\item{r}{Period of the pulse expresed in units of \code{1.5 * w} (so the pulses
don't overlap). Also controls the smoothness, along with \code{a}.}

\item{w}{Width of the midge pulse.}

\item{d}{Mid point of the first pulse in units of \code{w}.}

\item{tstep}{Step size in units of time. Defaults to \code{1}.}

\item{V}{Boolean for whether to include the V pool.}

\item{R}{Boolean for whether to include the R pool.
This pool isn't allowed when no V or H pool are present.}

\item{H}{Boolean for whether to include the H pool.}

\item{pool_starts}{Named list of initial values for each pool.
Names can include "N0", "D0", "P0", "V0", "H0", "R0", and "M0".
Any pools not included here will start at their equilibrium value, except
for midges that start with zero by default.}
}
\description{
For more info on the midge-pulse parameters (`a`, `b`, `r`, `w`, and `d`),
see `vignette("smooth_pulse", "mtf")`.
}
