% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_web.R
\name{plot_web}
\alias{plot_web}
\title{Plot output from the \code{food_web} function.}
\usage{
plot_web(web_df)
}
\arguments{
\item{web_df}{Dataframe output from the \code{food_web} function.}
}
\description{
Plot output from the \code{food_web} function.
}
