% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulses.R
\name{pulse_summary}
\alias{pulse_summary}
\title{Summarize aspects of the pulse of nitrogen content in a time series.}
\usage{
pulse_summary(x, measure, t0 = 0, p = NULL)
}
\arguments{
\item{x}{A vector of nitrogen content through time.}

\item{measure}{Character specifying which summary measure to return.
Options are as follows:
\describe{
    \item{aamp}{absolute amplitude of the pulse.}
    \item{ramp}{relative amplitude of the pulse.}
    \item{peakt}{when the pulse peak occurs.}
    \item{len}{how long the pulse lasts.}
    \item{skew}{ratio of time before to after the pulse peak.}
}}

\item{t0}{Number of time points to subtract from the time to pulse peak.
For example, if you want to ignore the first 10 time points, this would be
set to `10`. This only affects the output if `aamp == "peakt"`.
Defaults to `0`.}

\item{p}{Proportion of the total amplitude that is required for a time point to be
considered inside a pulse.
Only affects output (and is required) if `aamp %in% c("len", "skew")`.
Defaults to `NULL`, which will return an error if the above condition is
`TRUE`.}
}
\description{
Summarize aspects of the pulse of nitrogen content in a time series.
}
