% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulses.R
\name{test_midges}
\alias{test_midges}
\title{Test a midge pulse time series.}
\usage{
test_midges(tmax, a, b, r, w, d, tstep = 1)
}
\arguments{
\item{tmax}{Duration over which to run the model.}

\item{a}{Controls the smoothness of the pulse, along with \code{r}. If \code{a} is
sufficiently high, the pulse will always be rectangular.}

\item{b}{Maximum value of the midge pulse.}

\item{r}{Period of the pulse expresed in units of \code{1.5 * w} (so the pulses
don't overlap). Also controls the smoothness, along with \code{a}.}

\item{w}{Width of the midge pulse.}

\item{d}{Mid point of the first pulse in units of \code{w}.}

\item{tstep}{Step size in units of time. Defaults to \code{1}.}
}
\description{
Test a midge pulse time series.
}
