% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_web.R
\name{food_web}
\alias{food_web}
\title{Run a food web.}
\usage{
food_web(
  tmax,
  b,
  s,
  w,
  tstep = 1,
  pool_starts = NULL,
  ep_obj = NULL,
  other_pars = list(),
  midges_not_to = NULL
)
}
\arguments{
\item{tmax}{Duration over which to run the model.}

\item{b}{Maximum value of the midge pulse.}

\item{s}{Start time for midge pulse}

\item{w}{Width of the midge pulse.}

\item{tstep}{Step size in units of time. Defaults to \code{1}.}

\item{pool_starts}{Named list of initial values for each pool.
Names can include "N0", "D0", "P0", "V0", "H0", "R0", and "M0"
(for nitrogen, detritus, plant, detritivore, herbivore, predator, and midge
pools, respectively).
Any pools not included here will start at their equilibrium value, except
for midges that start at zero by default.}

\item{ep_obj}{Output from the \code{\link{equil_pools}} function that specifies
equilibrium pool sizes when you change variable values from defaults.
This is used when you pass something to \code{other_pars} that changes equilibrium
pool size(s).}

\item{other_pars}{A named list of other parameter values to use instead of defaults.
Possible parameter names include all column names in \code{par_estimates}
except the first three (see \code{?par_estimates} for a description of column names).}

\item{midges_not_to}{Character specifying whether midges should NOT go to detritus
or NOT to predators. If \code{"both"}, \code{NULL}, or \code{NA}, midges go to both.
Takes as input the following: \code{"both"}, \code{NULL}, \code{NA}, \code{"X"}, \code{"predator"},
\code{"D"}, or \code{"detritus"}. Defaults to \code{NULL}.}
}
\value{
A data frame with the following columns: \code{time}, \code{pool}, \code{N}.
}
\description{
Run a food web.
}
\examples{
# What happens to the food web bc of this midge pulse?
web_output <- food_web(tmax = 1000, b = 400, s = 10, w = 40)
web_output


}
