% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_web.R
\name{food_web}
\alias{food_web}
\title{Run a food web.}
\usage{
food_web(tmax, b, s, w,
         tstep = 1,
         .V = TRUE, .R = TRUE, .H = TRUE,
         .mM = 0.5, .aM = 1, .lM = 0.1,
         pool_starts = NULL)
}
\arguments{
\item{tmax}{Duration over which to run the model.}

\item{b}{Maximum value of the midge pulse.}

\item{s}{Start time for midge pulse}

\item{w}{Width of the midge pulse.}

\item{tstep}{Step size in units of time. Defaults to \code{1}.}

\item{.V}{Boolean for whether to include the V pool. Defaults to \code{TRUE}.}

\item{.R}{Boolean for whether to include the R pool. Defaults to \code{TRUE}.}

\item{.H}{Boolean for whether to include the H pool. Defaults to \code{TRUE}.}

\item{.iN}{Input rate from pool N. Options for this include \code{5}, \code{10}, or \code{20}.
Defaults to \code{10}.}

\item{.mM}{Loss rate of midges. Defaults to 0.01.}

\item{.aM}{Attack rate for midges. Defaults to 20.}

\item{.lM}{Loss rate of midges.}

\item{pool_starts}{Named list of initial values for each pool.
Names can include "N0", "D0", "P0", "V0", "H0", "R0", and "M0"
(for nitrogen, detritus, plant, detritivore, herbivore, predator, and midge
pools, respectively).
Any pools not included here will start at their equilibrium value, except
for midges that start at zero by default.}

\item{other_pars}{A named list of other parameter values to use instead of defaults.
Possible parameter names include all column names in \code{par_estimates}
except the first three (see \code{?par_estimates} for a description of column names).}
}
\value{
A data frame with the following columns: \code{time}, \code{pool}, \code{N}.
}
\description{
For more info on the midge-pulse parameters (\code{b}, \code{s}, \code{w}),
see \code{vignette("smooth_pulse", "mtf")}.
}
\examples{
# What happens to the food web bc of this midge pulse?
web_output <- food_web(tmax = 1000, a = 1e9, b = 0, r = 400, w = 40, d = 1)
plot_web(web_output)

# If you want to adjust starting value for predators:
web_output <- food_web(tmax = 1000, a = 1e9, b = 0, r = 400, w = 40, d = 1,
                       pool_starts = list(R0 = 10))
web_output

# If you want to remove predators from the food web entirely:
web_output <- food_web(tmax = 1000, a = 1e9, b = 0, r = 400, w = 40, d = 1,
                       .R = FALSE)
web_output

}
