% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_web.R, R/pulses.R
\name{food_web}
\alias{food_web}
\alias{test_midges}
\title{Run a food web.}
\usage{
food_web(tmax, a, b, r, w, d, tstep = 1, V = TRUE, R = TRUE,
  H = TRUE, pool_starts = NULL)

test_midges(tmax, a, b, r, w, d, tstep = 1)
}
\arguments{
\item{tmax}{Duration over which to run the model.}

\item{a}{Controls the smoothness of the pulse, along with \code{r}. If \code{a} is
sufficiently high, the pulse will always be rectangular.}

\item{b}{Maximum value of the midge pulse.}

\item{r}{Period of the pulse expresed in units of \code{1.5 * w} (so the pulses
don't overlap). Also controls the smoothness, along with \code{a}.}

\item{w}{Width of the midge pulse.}

\item{d}{Mid point of the first pulse in units of \code{w}.}

\item{tstep}{Step size in units of time. Defaults to \code{1}.}

\item{V}{Boolean for whether to include the V pool.}

\item{R}{Boolean for whether to include the R pool.
This pool isn't allowed when no V or H pool are present.}

\item{H}{Boolean for whether to include the H pool.}

\item{pool_starts}{Named list of initial values for each pool.
Names can include "N0", "D0", "P0", "V0", "H0", "R0", and "M0".
Any pools not included here will start at their equilibrium value, except
for midges that start with zero by default.}
}
\value{
A data frame with the following columns: \code{time}, \code{pool}, \code{N}.
}
\description{
For more info on the midge-pulse parameters (\code{a}, \code{b}, \code{r}, \code{w}, and \code{d}),
see \code{vignette("smooth_pulse", "mtf")}.
}
\section{Functions}{
\itemize{
\item \code{test_midges}: Test a midge pulse time series.
}}

\examples{
# Test a midge pulse:
midges <- test_midges(1000, a = 1e9, b = 10, r = 400, w = 40, d = 1)
plot(midges, type = 'l', ylab = "iM")

# What happens to the food web bc of this midge pulse?
web_output <- food_web(tmax = 1000, a = 1e9, b = 0, r = 400, w = 40, d = 1)
web_output

# If you want to adjust starting value for predators:
web_output <- food_web(tmax = 1000, a = 1e9, b = 0, r = 400, w = 40, d = 1,
                       pool_starts = list(R0 = 10))
web_output

# (Below is not yet coded bc we don't have the parameter estimates for anything except
# the full food web)
# # If you want to remove predators from the food web entirely:
# web_output <- food_web(tmax = 1000, a = 1e9, b = 0, r = 400, w = 40, d = 1,
                         R = FALSE)
# web_output

}
