% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class.R
\docType{data}
\name{web}
\alias{web}
\title{Foodweb class. (More documentation forthcoming...)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
web
}
\description{
Foodweb class. (More documentation forthcoming...)
}
\keyword{datasets}
